Separate TPG directory from wib code
