+++ include/fdreadoutlibs/tpg/RegisterToChannelNumber.hpp
