+++ src/tpg/DesignFIR.cpp
@@ -9,7 +9,7 @@
  * Licensing/copyright details are in the COPYING file that you should have
  * received with this code.
  */
-#include "fdreadoutlibs/wib/tpg/DesignFIR.hpp" // NOLINT(build/include)
+#include "fdreadoutlibs/tpg/DesignFIR.hpp" // NOLINT(build/include)
 
 #include <vector>
